+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Next version
 + Check if public IP is different from DNS record
++ Add browse zones test
 
 ## 0.2.2
 + Fix yaml